Finalized version.
